Change in signature of 'write_hexa' function, changing data type fom long to unsigned int

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jfidalgo <jfidalgo@student.42bar(...).com  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/22 15:19:42 by jfidalgo          #+#    #+#             */
-/*   Updated: 2024/02/28 15:51:05 by jfidalgo         ###   ########.fr       */
+/*   Updated: 2024/02/29 21:22:20 by jfidalgo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 int	write_char(int character);
 int	write_string(char *string);
 int	write_number(long number);
-int	write_hexa(long number, int uppercase);
+int	write_hexa(unsigned int number, int uppercase);
 int	write_pointer(unsigned long address);
 int	ft_printf(char const *format, ...);
